separated app into home + analytics screen
